backend/requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11

# Advanced NLP
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
scikit-learn==1.3.0
transformers==4.35.2
torch==2.1.1

# Utilities
python-dateutil==2.8.2
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
